Move upload beacon tests
