+++ test/src/tests-main.js
@@ -1,4 +1,4 @@
-import { MPConfig, workspaceToken} from './config';
+import { MPConfig, workspaceToken } from './config';
 
 var userApi = null;
 
@@ -7,8 +7,7 @@ beforeEach(function() {
     //mocha can't clean up after itself, so this lets
     //tests mock the current user and restores in between runs.
     if (!userApi) {
-        userApi = window.mParticle.getInstance().Identity
-            .getCurrentUser;
+        userApi = window.mParticle.getInstance().Identity.getCurrentUser;
     } else {
         window.mParticle.getInstance().Identity.getCurrentUser = userApi;
     }
@@ -20,7 +19,7 @@ beforeEach(function() {
         requestConfig: false,
         isDevelopmentMode: false,
     };
-    
+
     mParticle._resetForTests(MPConfig);
     delete mParticle._instances['default_instance'];
 });
@@ -28,6 +27,9 @@ beforeEach(function() {
 afterEach(function() {
     window.fetchMock.restore();
 });
+
+// NOTE: Beacon upload must come before core sdk tests beause of race condition
+import './tests-beaconUpload';
 
 import './tests-core-sdk';
 import './tests-temp-session-bug-fix';
